requirements.txt
# absl-py==2.4.0
# alembic==1.18.1
# annotated-doc==0.0.4
# annotated-types==0.7.0
# antlr4-python3-runtime==4.9.3
# anyio==4.12.1
# asttokens==3.0.1
# blinker==1.9.0
# cachetools==6.2.4
# certifi==2026.1.4
# cffi==2.0.0
# charset-normalizer==3.4.4
# click==8.3.1
# cloudpickle==3.1.2
# colorama==0.4.6
# comm==0.2.3
# contourpy==1.3.3
# cryptography==46.0.3
# cycler==0.12.1
# databricks-sdk==0.78.0
# debugpy==1.8.19
# decorator==5.2.1
# docker==7.1.0
# executing==2.2.1
# fastapi==0.128.0
# filelock==3.20.3
# Flask==3.1.2
# flask-cors==6.0.2
# fonttools==4.61.1
# fsspec==2026.1.0
# gitdb==4.0.12
# GitPython==3.1.46
# google-auth==2.47.0
# graphene==3.4.3
# graphql-core==3.2.7
# graphql-relay==3.2.0
# greenlet==3.3.0
# h11==0.16.0
# hf-xet==1.2.0
# httpcore==1.0.9
# httpx==0.28.1
# huey==2.6.0
# huggingface_hub==1.3.2
# idna==3.11
# importlib_metadata==8.7.1
# ipykernel==7.1.0
# ipython==9.9.0
# ipython_pygments_lexers==1.1.1
# itsdangerous==2.2.0
# jedi==0.19.2
# Jinja2==3.1.6
# joblib==1.5.3
# jupyter_client==8.8.0
# jupyter_core==5.9.1
# kiwisolver==1.4.9
# Mako==1.3.10
# MarkupSafe==3.0.3
# matplotlib==3.10.8
# matplotlib-inline==0.2.1
# ml_collections==1.1.0
# mlflow==3.8.1
# mlflow-skinny==3.8.1
# mlflow-tracing==3.8.1
# mpmath==1.3.0
# nest-asyncio==1.6.0
# networkx==3.6.1
# numpy==2.4.1
# omegaconf==2.3.0
# opencv-python==4.13.0.90
# opentelemetry-api==1.39.1
# opentelemetry-proto==1.39.1
# opentelemetry-sdk==1.39.1
# opentelemetry-semantic-conventions==0.60b1
# packaging==25.0
# pandas==2.3.3
# parso==0.8.5
# pillow==12.1.0
# platformdirs==4.5.1
# prompt_toolkit==3.0.52
# protobuf==6.33.4
# psutil==7.2.1
# pure_eval==0.2.3
# pyarrow==22.0.0
# pyasn1==0.6.2
# pyasn1_modules==0.4.2
# pycparser==2.23
# pydantic==2.12.5
# pydantic_core==2.41.5
# Pygments==2.19.2
# pyparsing==3.3.1
# python-dateutil==2.9.0.post0
# python-dotenv==1.2.1
# pytz==2025.2
# pywin32==311
# PyYAML==6.0.3
# pyzmq==27.1.0
# requests==2.32.5
# rsa==4.9.1
# safetensors==0.7.0
# scikit-learn==1.8.0
# scipy==1.17.0
# seaborn==0.13.2
# setuptools==80.9.0
# shellingham==1.5.4
# six==1.17.0
# smmap==5.0.2
# SQLAlchemy==2.0.45
# sqlparse==0.5.5
# stack-data==0.6.3
# starlette==0.50.0
# sympy==1.14.0
# thop==0.1.1.post2209072238
# threadpoolctl==3.6.0
# timm==1.0.24
# torch==2.9.1
# torchaudio==2.9.1
# torchvision==0.24.1
# tornado==6.5.4
# tqdm==4.67.3
# traitlets==5.14.3
# typer-slim==0.21.1
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# tzdata==2025.3
# urllib3==2.6.3
# uvicorn==0.40.0
# waitress==3.0.2
# wcwidth==0.3.0
# Werkzeug==3.1.5
# zipp==3.23.0


torch>=2.0
torchvision>=0.15
numpy>=1.24
PyYAML>=6.0
omegaconf>=2.3
Pillow>=10.0
scipy>=1.10
opencv-python>=4.8
tqdm>=4.65
mlflow>=2.0
matplotlib>=3.7